+++ changelog.md
@@ -41,6 +41,7 @@ Codion Change Log
 - EntityConnectionProvider.clientType() and clientVersion() now return Optional.
 ### is.codion.framework.model
 - DefaultForeignKeyDetailModelLink now uses TableConditionModel.persist() to exclude the associated foreign key condition from being cleared.
+- EntityEditModel.EditableEntity and EditableValue renamed EntityEditor and ValueEditor respectively.
 ### is.codion.swing.framework.ui
 - TabbedApplicationLayout.applicationPanel() removed.
 - EntityApplicationPanel.displayEntityPanelFrame() and displayEntityPanelDialog() now initialize the focus by activating the panel when opened.